Revert "try splitting tests in different files"
This reverts commit 4845b80026424c00b105f8f35483c8d445a6b1b9.
